Fix to ensure splash page is properly overwritten on update. Review URL: https://codereview.chromium.org//11038068
git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@13312 260f80e4-7a28-3924-810f-c04153c831b5

+++ dart/editor/tools/plugins/com.google.dart.tools.update.core/src/com/google/dart/tools/update/core/internal/UpdateUtils.java
@@ -167,11 +167,16 @@ public class UpdateUtils {
 
   /**
    * Delete the given file. If the file is a directory, recurse and delete contents.
+   * <p>
+   * If the file does not exist, ignore.
    * 
    * @param file the file to delete
    * @param monitor the progress monitor
    */
   public static void delete(File file, IProgressMonitor monitor) {
+    if (!file.exists()) {
+      return;
+    }
     if (file.isFile()) {
       file.delete();
       monitor.worked(1);
